Extract 'MoreMessages' to a separate file
